Organize CLI command helpers
