+++ src/cli/mod.rs
@@ -2,17 +2,8 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 include!("prelude.rs");
-include!("workflow_output.rs");
 include!("transport.rs");
 include!("resource_targets.rs");
-include!("resource_output.rs");
-include!("knowledge_io.rs");
-include!("workflow_api.rs");
-include!("knowledge_sync.rs");
-include!("resource_get.rs");
-include!("resource_get_tests.rs");
-include!("resource_delete.rs");
 include!("entry.rs");
 include!("manifest_render.rs");
-include!("sdk_gen.rs");
 include!("cli_render_tests.rs");
