+++ src/cli/prelude.rs
@@ -1,23 +1,11 @@
 // Copyright (C) 2026 Impala Systems, Inc.
 // SPDX-License-Identifier: AGPL-3.0-only
 
-use crate::control::resource_model::{self, ChannelSubscriptionResourceExt, TypedResource};
-use crate::gateway::rpc::manifests::{Knowledge, KnowledgeSpec, ObjectMeta};
-use crate::gateway::rpc::proto::gateway_service_client::GatewayServiceClient;
-use crate::gateway::rpc::proto::{
-    CancelWorkflowRunRequest, CreateResourceRequest, CreateWorkflowRunRequest,
-    DeleteResourceRequest, GetResourceRequest, GetWorkflowRunRequest, ListNamespacesRequest,
-    ListResourcesRequest, ListWorkflowRunsRequest, ResumeWorkflowRunRequest,
-    StreamWorkflowEventsRequest,
-};
-use crate::gateway::rpc::{data_proto, resources_proto};
 use anyhow::{Context, Result};
 use clap::Parser;
-use futures::StreamExt;
 use minijinja::{context, Environment, UndefinedBehavior};
 use reqwest::header::CONTENT_TYPE;
 use serde::Deserialize;
-use serde_json::json;
 use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
